Add changes to test names

+++ src/test/spec/recipe.spec.ts
@@ -6,24 +6,27 @@ import { generateToken, deleteAllUsers } from '../mocks/user.mocks';
 
 let token = '';
 let id = '';
+const HEADER_AUTHORIZATION = 'Authorization';
 
 beforeEach('Add new user and return token-Recipe', async function () {
-  await deleteAllRecipes();
-  await deleteAllUsers();
-  const data = await generateToken();
-  const _id = data._id;
-  id = await createRecipeTest(_id);
-  token = `Bearer ${data.token}`;
-});
-
-afterEach('Delete all recipes and all users', async function () {
-  await deleteAllRecipes();
+  if (token === '') {
+    await deleteAllRecipes();
+    await deleteAllUsers();
+    const data = await generateToken();
+    const _id = data._id;
+    id = await createRecipeTest(_id);
+    token = `Bearer ${data.token}`;
+  }
 });
 
 describe('Recipe', function () {
   describe('POST /recipes/', function () {
     it('should create recipe', async function () {
-      const response = await chai.request(app).post('/recipes').set('Authorization', token).send(createRecipePayload);
+      const response = await chai
+        .request(app)
+        .post('/recipes')
+        .set(HEADER_AUTHORIZATION, token)
+        .send(createRecipePayload);
 
       expect(response).to.have.status(StatusCodes.CREATED);
       expect(response.error).to.be.false;
@@ -33,7 +36,7 @@ describe('Recipe', function () {
 
   describe('GET /recipes/', function () {
     it('should get recipes', async function () {
-      const response = await chai.request(app).get('/recipes').set('Authorization', token);
+      const response = await chai.request(app).get('/recipes').set(HEADER_AUTHORIZATION, token);
 
       expect(response).to.have.status(StatusCodes.OK);
       expect(response.error).to.be.false;
@@ -43,19 +46,16 @@ describe('Recipe', function () {
 
   describe('GET /recipes/all', function () {
     it('should get all recipes', async function () {
-      const response = await chai.request(app).get('/recipes/all').set('Authorization', token);
+      const response = await chai.request(app).get('/recipes/all').set(HEADER_AUTHORIZATION, token);
 
       expect(response).to.have.status(StatusCodes.OK);
       expect(response.error).to.be.false;
       expect(response.body).to.not.be.null;
     });
   });
-  describe('GET /recipes/', function () {
+  describe('GET /recipes/:recipeId', function () {
     it('should get recipe by id', async function () {
-      const response = await chai
-        .request(app)
-        .get('/recipes/' + id)
-        .set('Authorization', token);
+      const response = await chai.request(app).get(`/recipes/${id}`).set(HEADER_AUTHORIZATION, token);
 
       expect(response).to.have.status(StatusCodes.OK);
       expect(response.error).to.be.false;
@@ -63,12 +63,12 @@ describe('Recipe', function () {
     });
   });
 
-  describe('PUT/recipes/', function () {
-    it('should put recipe by id', async function () {
+  describe('PUT/recipes/:recipeId', function () {
+    it('should return code 200', async function () {
       const response = await chai
         .request(app)
-        .put('/recipes/' + id)
-        .set('Authorization', token)
+        .put(`/recipes/${id}`)
+        .set(HEADER_AUTHORIZATION, token)
         .send(createRecipePayload);
 
       expect(response).to.have.status(StatusCodes.OK);
@@ -77,16 +77,18 @@ describe('Recipe', function () {
     });
   });
 
-  describe('DELETE/recipes/', function () {
-    it('should delete recipe by id', async function () {
-      const response = await chai
-        .request(app)
-        .delete('/recipes/' + id)
-        .set('Authorization', token);
+  describe('DELETE/recipes/:recipeId', function () {
+    it('should return code 204', async function () {
+      const response = await chai.request(app).delete(`/recipes/${id}`).set(HEADER_AUTHORIZATION, token);
 
       expect(response).to.have.status(StatusCodes.NO_CONTENT);
       expect(response.error).to.be.false;
       expect(response.body).to.not.be.null;
+    });
+    it('should return code 404', async function () {
+      const response = await chai.request(app).delete(`/recipes/${id}`).set(HEADER_AUTHORIZATION, token);
+
+      expect(response).to.have.status(StatusCodes.NOT_FOUND);
     });
   });
 });
